requirements.txt
pandas
numpy
scikit-learn
joblib
kafka-python
